refactor: 전략 패키지 이동
