+++ src/main/java/wooteco/subway/path/application/PathService.java
@@ -3,9 +3,8 @@ package wooteco.subway.path.application;
 import org.springframework.stereotype.Service;
 import wooteco.subway.line.dao.LineDao;
 import wooteco.subway.line.domain.Line;
-import wooteco.subway.path.domain.DijkstraShortestPathStrategy;
+import wooteco.subway.path.domain.strategy.shortestpath.DijkstraShortestPathStrategy;
 import wooteco.subway.path.domain.Path;
-import wooteco.subway.path.domain.ShortestPathStrategy;
 import wooteco.subway.path.dto.PathResponse;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationResponse;
