+++ src/main/java/wooteco/subway/path/domain/strategy/shortestpath/DijkstraShortestPathStrategy.java
@@ -1,16 +1,16 @@
-package wooteco.subway.path.domain;
+package wooteco.subway.path.domain.strategy.shortestpath;
 
 import org.jgrapht.alg.interfaces.ShortestPathAlgorithm;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
-import org.jgrapht.alg.shortestpath.FloydWarshallShortestPaths;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import wooteco.subway.line.domain.Line;
+import wooteco.subway.path.domain.PathGraph;
 import wooteco.subway.station.domain.Station;
 
 import java.util.List;
 
-public class FloydWarshallShortestPathStrategy {
+public class DijkstraShortestPathStrategy extends ShortestPathStrategy {
     public ShortestPathAlgorithm<Station, DefaultWeightedEdge> match(List<Line> lines) {
-        return new FloydWarshallShortestPaths<>(new PathGraph(lines).graph());
+        return new DijkstraShortestPath<>(new PathGraph(lines).graph());
     }
 }
