+++ src/main/java/wooteco/subway/path/domain/strategy/shortestpath/ShortestPathStrategy.java
@@ -1,8 +1,6 @@
-package wooteco.subway.path.domain;
+package wooteco.subway.path.domain.strategy.shortestpath;
 
 import org.jgrapht.alg.interfaces.ShortestPathAlgorithm;
-import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
-import org.jgrapht.alg.shortestpath.FloydWarshallShortestPaths;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import wooteco.subway.line.domain.Line;
 import wooteco.subway.station.domain.Station;
